// web/lib/api/generated/model/stockAlertsListStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export type StockAlertsListStatus = typeof StockAlertsListStatus[keyof typeof StockAlertsListStatus];


export const StockAlertsListStatus = {
  '': '',
  OPEN: 'OPEN',
  RESOLVED: 'RESOLVED',
} as const;