// web/lib/api/generated/model/stockAlertsListType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export type StockAlertsListType = typeof StockAlertsListType[keyof typeof StockAlertsListType];


export const StockAlertsListType = {
  BACKORDER_DEMAND: 'BACKORDER_DEMAND',
  LOW_STOCK: 'LOW_STOCK',
  OUT_OF_STOCK: 'OUT_OF_STOCK',
} as const;